added unfinished pong_conv agent
This agent was originally from a branch of anyrl. I am trying to get a network to learn to play Pong using the raw (unscaled & color) images of the screen. This implementation will be a hacky mess until I get it working, at which point I will clean it up.